api/requirements.txt
Flask>=2.2
gunicorn>=20.1.0
Jinja2==3.0.1
itsdangerous==2.0.1
Werkzeug>=2.2.2
requests==2.28.2
flask-sqlalchemy==3.0.2
certifi==2022.12.7
marshmallow>=3.12.1
pika==1.3.1
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
ConfigParser==5.3.0
psycopg2-binary>=0.0
google-cloud-storage==2.8.0
google-cloud-pubsub==2.16.1